# contracts/openapi/api-v1.json
{
  "components": {
    "schemas": {
...
      "Body_upload_op_api_v1_parametrization_op_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_op_api_v1_parametrization_op_upload_post",
        "type": "object"
      },
...
      "CertificateVerificationResultV1": {
        "additionalProperties": false,
        "description": "Result returned by ``POST /certification/verify/{certificate_id}``.",
        "properties": {
          "certificate_id": {
            "title": "Certificate Id",
            "type": "string"
          },
          "checked_modules": {
            "items": {
              "type": "string"
            },
            "title": "Checked Modules",
            "type": "array"
          },
          "drift": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Drift",
            "type": "object"
          },
          "simulation_id": {
            "title": "Simulation Id",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "certificate_id",
          "simulation_id",
          "valid"
        ],
        "title": "CertificateVerificationResultV1",
        "type": "object"
      },
...
      "ExecutionCertificateV1": {
        "additionalProperties": false,
        "description": "Public DTO for an ``ExecutionCertificate``.",
        "properties": {
          "baseline_matched": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Matched"
          },
          "certificate_id": {
            "title": "Certificate Id",
            "type": "string"
          },
          "issued_at": {
            "title": "Issued At",
            "type": "string"
          },
          "lineage_hash": {
            "title": "Lineage Hash",
            "type": "string"
          },
          "request_hash": {
            "title": "Request Hash",
            "type": "string"
          },
          "result_hash": {
            "title": "Result Hash",
            "type": "string"
          },
          "simulation_id": {
            "title": "Simulation Id",
            "type": "string"
          },
          "validation_results": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Validation Results",
            "type": "object"
          },
          "version_metadata": {
            "additionalProperties": true,
            "title": "Version Metadata",
            "type": "object"
          }
        },
        "required": [
          "certificate_id",
          "simulation_id",
          "issued_at",
          "request_hash",
          "result_hash",
          "lineage_hash"
        ],
        "title": "ExecutionCertificateV1",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Motor de pricing + API de parametrización y simulación",
    "title": "NEXA Simulator API",
    "version": "1.0.0",
    "x-contract-version": "api-v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/certification/certificate/{certificate_id}": {
      "get": {
        "operationId": "get_certificate_api_v1_certification_certificate__certificate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "certificate_id",
            "required": true,
            "schema": {
              "title": "Certificate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionCertificateV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Certificate",
        "tags": [
          "certification"
        ]
      }
    },
    "/api/v1/certification/certificates": {
      "get": {
        "description": "List most recent certificates (descending by mtime).",
        "operationId": "list_certificates_api_v1_certification_certificates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExecutionCertificateV1"
                  },
                  "title": "Response List Certificates Api V1 Certification Certificates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Certificates",
        "tags": [
          "certification"
        ]
      }
    },
    "/api/v1/certification/verify/{certificate_id}": {
      "post": {
        "description": "Re-validate ``certificate_id`` against the live parametrization hashes.",
        "operationId": "verify_certificate_api_v1_certification_verify__certificate_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "certificate_id",
            "required": true,
            "schema": {
              "title": "Certificate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVerificationResultV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Certificate not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Certificate",
        "tags": [
          "certification"
        ]
      }
    },
...
    "/api/v1/parametrization/op/upload": {
      "post": {
        "operationId": "upload_op_api_v1_parametrization_op_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_op_api_v1_parametrization_op_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Op",
        "tags": [
          "parametrization-op"
        ]
      }
    },
...
    "/api/v1/parametrization/op/{version_id}": {
      "delete": {
        "operationId": "delete_op_api_v1_parametrization_op__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Op",
        "tags": [
          "parametrization-op"
        ]
      }
    },
    "/api/v1/parametrization/op/{version_id}/activate": {
      "get": {
        "operationId": "activate_op_api_v1_parametrization_op__version_id__activate_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Op",
        "tags": [
          "parametrization-op"
        ]
      }
    },
...
  }
}